Changed long into to int

+++ 0x03-more_functions_nested_loops/2-mul.c
@@ -9,7 +9,7 @@
 
 int mul(int a, int b)
 {
-	long int ab = (a * b);
+	int ab = (a * b);
 
 	return (ab);
 }
